+++ frontend/components/builder/elements/map/Props.ts
@@ -1,0 +1,4 @@
+export interface MapProps {
+    routeId: number,
+    segmentsIds: number[],
+}