+++ frontend/types/color.ts
@@ -13,17 +13,33 @@ export type Color = RGB | RGBA | HEX;
  */
 export function addAlphaToColor(color: Color, alpha: number): string {
     // Regular expressions to match `rgb` and `rgba` formats
+    const hexShorthandRegex = /^#?([a-f\d])([a-f\d])([a-f\d])$/i;
+    const hexRegex = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i;
     const rgbRegex = /^rgb\((\d{1,3}),\s*(\d{1,3}),\s*(\d{1,3})\)$/;
     const rgbaRegex = /^rgba\((\d{1,3}),\s*(\d{1,3}),\s*(\d{1,3}),\s*(0|1|0?\.\d+)\)$/;
 
-    // Check for `rgb` format
+    const hexShorthandMatch = color.match(hexShorthandRegex);
+    if (hexShorthandMatch) {
+        const r = parseInt(hexShorthandMatch[1] + hexShorthandMatch[1], 16);
+        const g = parseInt(hexShorthandMatch[2] + hexShorthandMatch[2], 16);
+        const b = parseInt(hexShorthandMatch[3] + hexShorthandMatch[3], 16);
+        return `rgba(${r}, ${g}, ${b}, ${alpha})`;
+    }
+
+    const hexRegexMatch = color.match(hexRegex);
+    if (hexRegexMatch) {
+        const r = parseInt(hexRegexMatch[1], 16);
+        const g = parseInt(hexRegexMatch[2], 16);
+        const b = parseInt(hexRegexMatch[3], 16);
+        return `rgba(${r}, ${g}, ${b}, ${alpha})`;
+    }
+
     const rgbMatch = color.match(rgbRegex);
     if (rgbMatch) {
         const [, r, g, b] = rgbMatch;
         return `rgba(${r}, ${g}, ${b}, ${alpha})`;
     }
 
-    // Check for `rgba` format
     const rgbaMatch = color.match(rgbaRegex);
     if (rgbaMatch) {
         const [, r, g, b] = rgbaMatch;